Fix AVPlayer race condition
A reported bug described a button flashing without the corresponding
note sounding. It seemed like this may have been due to the player
being asked to sound the note ("play") prior to the note's underlying
asset -- aka, the sound resource -- being loaded. Each note only needs
to be loaded once, but we can't play it and expect it to sound until
it has been loaded.

Additionally, it seems like I was doing too much. There only needs to
be one player created. I took all that crud out.

WARNING: As with any race-condition, there is no real test. It seems
to work though, at the time of this writing.